requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
datasets==4.6.1
debugpy==1.8.20
decorator==5.2.1
dill==0.4.0
executing==2.2.1
filelock==3.25.0
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ipykernel==7.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
peft>=0.12.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.23.0
